Fix bug in mobility-based get_region_populations.
Previously, when no date was specified, for each region we were returning the
latest available data available (so we could have data for multiple different
dates). Now, we find the most recent date for which we have data, and return
data for all regions that have data for that date.